chore: changes after cr
