+++ src/components/composite/DustFlow/constants.ts
@@ -145,8 +145,6 @@ export const multicallAbi = [
   },
 ] as const;
 
-export const MULTICALL3_ADDRESS = '0xca11bde05977b3631167028862be2a173976ca11';
-
 export const APPROVE_ABI = [
   {
     inputs: [
