+++ src/components/composite/JumperWidget/types.ts
@@ -1,4 +1,4 @@
-import { createElement, type ReactNode } from 'react';
+import { type ReactNode } from 'react';
 import type { z } from 'zod';
 import type { StatusBottomSheetProps } from '@/components/composite/StatusBottomSheet/StatusBottomSheet';
