+++ src/utils/isNil.ts
@@ -1,0 +1,2 @@
+export const isNil = (value: unknown): value is null | undefined =>
+  value === null || value === undefined;
